pyproject.toml
[project]
name = "john-atreides-bot"
version = "0.1.0"
description = "This is a Discord bot for mostly personal use, it's supposed to keep track of your Dune: Awakening base fuel/fees uptime."
authors = [
    {name = "Denis Verkhoturov",email = "sainte66@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
